This PR is connected with #6411 
Discarding elements from generators takes up most of the working time in `InitSampling`.
Since stdlibc++ doesn't have any random engines with o(n) complexity (little-o), we only can optimize the number of discarded elements.
 `std::bernoulli_distribution` requires 64-bit input, so in previous version we had to discard twice as much elements as now.
This little optimization gives us ~2.12 speed up of `InitSampling` time, which translates into up to 16% speed up of the whole training time when `subsampling < 1`
The quality remains the same:


Mortgage dataset |  |  |
-- | -- | --
version | training time | RMSE
original | 35.479 | 0.009271
optimized | 30.538 | 0.009262


Santander dataset |   |   |  | |
-- | -- | -- | -- | --
version | training time, s | InitData time, s | init improvement, % | Log Loss
original | 249.196 | 28.784 | 0.00 | 0.16607
optimized | 239.298 | 17.173 | 67.61 | 0.16610
no discard | 224.347 | 11.221 | 156.51 | 0.16613


Higgs dataset |  |  |  |  |
-- | -- | -- | -- | --
version | training time, s | InitData time, s | init improvement, % | Log Loss
original | 34.235 | 14.373 | 0.00 | 0.09339
optimized | 29.201 | 8.316 | 72.84 | 0.09409